Even more mods

+++ Assets/Scripts/Bullet.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Bullet : MonoBehaviour {
 
-    public int damage = 8;
+    public float damage = 8;
 
     void OnTriggerEnter2D(Collider2D other) {
         if (other.gameObject.tag == "Wall") {
